requirements.txt
django==1.6
pyechonest==8.0.1
iron_celery==0.4.0
django-celery==3.1.1
django-graphos==0.0.2a0
musicbrainzngs==0.4
django-autocomplete-light==2.0.0a8
south==0.8.4
